feat: rework init flow
